\name{AICcmodavg-defunct}
\Rdversion{1.1}
\alias{AICcmodavg-defunct}
\alias{AICc.mult}
\alias{AICc.unmarked}
\alias{extract.LL}
\alias{extract.LL.coxph}
\alias{extract.LL.unmarked}
\alias{aictab.clm}
\alias{aictab.clmm}
\alias{aictab.coxph}
\alias{aictab.glm}
\alias{aictab.gls}
\alias{aictab.lm}
\alias{aictab.lme}
\alias{aictab.mer}
\alias{aictab.merMod}
\alias{aictab.mult}
\alias{aictab.nlme}
\alias{aictab.nls}
\alias{aictab.polr}
\alias{aictab.rlm}
\alias{aictab.unmarked}
\alias{dictab.bugs}
\alias{dictab.rjags}
\alias{modavg.clm}
\alias{modavg.clmm}
\alias{modavg.coxph}
\alias{modavg.glm}
\alias{modavg.gls}
\alias{modavg.lme}
\alias{modavg.clm}
\alias{modavg.mer}
\alias{modavg.merMod}
\alias{modavg.mult}
\alias{modavg.polr}
\alias{modavg.rlm}
\alias{modavg.unmarked}
\alias{modavg.effect}
\alias{modavg.effect.glm}
\alias{modavg.effect.gls}
\alias{modavg.effect.lme}
\alias{modavg.effect.mer}
\alias{modavg.effect.merMod}
\alias{modavg.effect.rlm}
\alias{modavg.effect.unmarked}
\alias{modavg.shrink}
\alias{modavg.shrink.clm}
\alias{modavg.shrink.clmm}
\alias{modavg.shrink.coxph}
\alias{modavg.shrink.glm}
\alias{modavg.shrink.gls}
\alias{modavg.shrink.lme}
\alias{modavg.shrink.mer}
\alias{modavg.shrink.merMod}
\alias{modavg.shrink.mult}
\alias{modavg.shrink.polr}
\alias{modavg.shrink.rlm}
\alias{modavg.shrink.unmarked}
\alias{modavgpred}
\alias{modavgpred.glm}
\alias{modavgpred.gls}
\alias{modavgpred.lme}
\alias{modavgpred.mer}
\alias{modavgpred.merMod}
\alias{modavgpred.rlm}
\alias{modavgpred.unmarked}
\alias{mult.comp}
\alias{predictSE.zip}

\title{
Defunct Functions in AICcmodavg Package
}
\description{
The functions listed below have been removed from the
\code{AICcmodavg} package. 
}
\usage{
AICc.mult(\dots)
AICc.unmarked(\dots)
extract.LL(\dots)
extract.LL.coxph(\dots)
extract.LL.unmarked(\dots)
aictab.clm(\dots)
aictab.clmm(\dots)
aictab.coxph(\dots)
aictab.glm(\dots)
aictab.gls(\dots)
aictab.lm(\dots)
aictab.lme(\dots)
aictab.mer(\dots)
aictab.merMod(\dots)
aictab.mult(\dots)
aictab.nlme(\dots)
aictab.nls(\dots)
aictab.polr(\dots)
aictab.rlm(\dots)
aictab.unmarked(\dots)
dictab.bugs(\dots)
dictab.rjags(\dots)
modavg.clm(\dots)
modavg.clmm(\dots)
modavg.coxph(\dots)
modavg.glm(\dots)
modavg.gls(\dots)
modavg.lme(\dots)
modavg.mer(\dots)
modavg.merMod(\dots)
modavg.mult(\dots)
modavg.polr(\dots)
modavg.rlm(\dots)
modavg.unmarked(\dots)
modavg.effect(\dots)
modavg.effect.glm(\dots)
modavg.effect.gls(\dots)
modavg.effect.lme(\dots)
modavg.effect.mer(\dots)
modavg.effect.merMod(\dots)
modavg.effect.rlm(\dots)
modavg.effect.unmarked(\dots)
modavg.shrink(\dots)
modavg.shrink.clm(\dots)
modavg.shrink.clmm(\dots)
modavg.shrink.coxph(\dots)
modavg.shrink.glm(\dots)
modavg.shrink.gls(\dots)
modavg.shrink.lme(\dots)
modavg.shrink.mer(\dots)
modavg.shrink.merMod(\dots)
modavg.shrink.mult(\dots)
modavg.shrink.polr(\dots)
modavg.shrink.rlm(\dots)
modavg.shrink.unmarked(\dots)
modavgpred(\dots)
modavgpred.glm(\dots)
modavgpred.gls(\dots)
modavgpred.lme(\dots)
modavgpred.mer(\dots)
modavgpred.merMod(\dots)
modavgpred.rlm(\dots)
modavgpred.unmarked(\dots)
mult.comp(\dots)
predictSE.zip(\dots)
}

\arguments{
  \item{\dots}{
    arguments passed to the function.
  }
}
\details{
  \code{AICc.mult} has been replaced by \code{AICc.multinom}.

  \code{AICc.unmarked} has been replaced by \code{AICc.unmarkedFit}.

  \code{extract.LL} has been replaced by \code{extractLL}.
    
  \code{extract.LL.coxph} has been replaced by \code{extractLL.coxph}.

  \code{extract.LL.unmarked} has been replaced by
  \code{extractLL.unmarkedFit}.

  \code{aictab.clm} has been replaced by \code{aictab.AICsclm.clm}. 
  
  \code{aictab.clmm} has been replaced by \code{aictab.AICclmm}.
  
  \code{aictab.coxph} has been replaced by \code{aictab.AICcoxph}. 
  
  \code{aictab.glm} has been replaced by \code{aictab.AICglm.lm}.
  
  \code{aictab.gls} has been replaced by \code{aictab.AICgls}.
  
  \code{aictab.lm} has been replaced by \code{aictab.AIClm}.
  
  \code{aictab.lme} has been replaced by \code{aictab.AIClme}.
  
  \code{aictab.mer} has been replaced by \code{aictab.AICmer}.
  
  \code{aictab.merMod} has been replaced by \code{aictab.AIClmerMod},
  \code{aictab.AICglmerMod}, or \code{aictab.AICnlmerMod},
  depending on the class of the objects. 
  
  \code{aictab.mult} has been replaced by
  \code{aictab.AICmultinom.nnet}.
  
  \code{aictab.nlme} has been replaced by \code{aictab.AICnlme}. 
  
  \code{aictab.nls} has been replaced by \code{aictab.AICnls}.
  
  \code{aictab.polr} has been replaced by \code{aictab.AICpolr}. 
  
  \code{aictab.rlm} has been replaced by \code{aictab.AICrlm.lm}. 
  
  \code{aictab.unmarked} has been replaced by
  \code{aictab.AICunmarkedFitOccu},
  \code{aictab.AICunmarkedFitColExt},
  \code{aictab.AICunmarkedFitOccuRN},
  \code{aictab.AICunmarkedFitPCount},
  \code{aictab.AICunmarkedFitPCO}, \code{aictab.AICunmarkedFitDS},
  \code{aictab.AICunmarkedFitGDS},
  \code{aictab.AICunmarkedFitOccuFP},
  \code{aictab.AICunmarkedFitMPois},
  \code{aictab.AICunmarkedFitGMM}, or
  \code{aictab.AICunmarkedFitGPC}, depending on the class of the
  objects.

  \code{dictab.bugs} has been replaced by \code{dictab.AICbugs}. 

  \code{dictab.jags} has been replaced by \code{dictab.AICjags}. 

  \code{modavg.clm} has been replaced by \code{modavg.AICsclm.clm}.
  
  \code{modavg.clmm} has been replaced by \code{modavg.AICsclm.clm}.

  \code{modavg.coxph} has been replaced by \code{modavg.AICcoxph}.
  
  \code{modavg.glm} has been replaced by \code{modavg.AIClm} or
  \code{modavg.AICglm.lm}, depending on the class of the objects.
  
  \code{modavg.gls} has been replaced by \code{modavg.AICgls}.

  \code{modavg.lme} has been replaced by \code{modavg.AIClme}.
  
  \code{modavg.mer} has been replaced by \code{modavg.AICmer}.
  
  \code{modavg.merMod} has been replaced by \code{modavg.AIClmerMod} or
  \code{modavg.AICglmerMod}, depending on the class of the objects.
  
  \code{modavg.mult} has been replaced by \code{modavg.AICmultinom.nnet}.
  
  \code{modavg.polr} has been replaced by \code{modavg.AICpolr}.
  
  \code{modavg.rlm} has been replaced by \code{modavg.AICrlm.lm}.
  
  \code{modavg.unmarked} has been replaced by
  \code{modavg.AICunmarkedFitOccu},
  \code{modavg.AICunmarkedFitColExt},
  \code{modavg.AICunmarkedFitOccuRN},
  \code{modavg.AICunmarkedFitPCount},
  \code{modavg.AICunmarkedFitPCO}, \code{modavg.AICunmarkedFitDS},
  \code{modavg.AICunmarkedFitGDS},
  \code{modavg.AICunmarkedFitOccuFP},
  \code{modavg.AICunmarkedFitMPois},
  \code{modavg.AICunmarkedFitGMM}, or
  \code{modavg.AICunmarkedFitGPC}, depending on the class of the
  objects.

  \code{modavg.effect} has been replaced by \code{modavgEffect}.
  
  \code{modavg.effect.glm} has been replaced by
  \code{modavgEffect.AICglm.lm} or \code{modavgEffect.AIClm},
  depending on the class of the objects.

  \code{modavg.effect.gls} has been replaced by
  \code{modavgEffect.AICgls}.

  \code{modavg.effect.lme} has been replaced by
  \code{modavgEffect.AIClme}. 

\code{modavg.effect.mer} has been replaced by
\code{modavgEffect.AICmer}. 

\code{modavg.effect.merMod} has been replaced by
\code{modavgEffect.AICglmerMod} or
\code{modavgEffect.AIClmerMod}, depending on the class of the objects.

\code{modavg.effect.rlm} has been replaced by
\code{modavgEffect.AICrlm.lm}.

\code{modavg.effect.unmarked} has been replaced by
  \code{modavgEffect.AICunmarkedFitOccu}, 
  \code{modavgEffect.AICunmarkedFitColExt},
  \code{modavgEffect.AICunmarkedFitOccuRN},
  \code{modavgEffect.AICunmarkedFitPCount},
  \code{modavgEffect.AICunmarkedFitPCO},
  \code{modavgEffect.AICunmarkedFitDS}, 
  \code{modavgEffect.AICunmarkedFitGDS},
  \code{modavgEffect.AICunmarkedFitOccuFP},
  \code{modavgEffect.AICunmarkedFitMPois},
  \code{modavgEffect.AICunmarkedFitGMM}, or
  \code{modavgEffect.AICunmarkedFitGPC}, depending on the class of the
  objects.

  \code{modavg.shrink} has been replaced by \code{modavgShrink}. 
    
  \code{modavg.shrink.clm} has been replaced by
  \code{modavgShrink.AICsclm.clm}. 

  \code{modavg.shrink.clmm} has been replaced by
  \code{modavgShrink.AICclmm}. 

  \code{modavg.shrink.coxph} has been replaced by
  \code{modavgShrink.AICcoxph}. 

  \code{modavg.shrink.glm} has been replaced by
  \code{modavgShrink.AICglm.lm} or \code{modavgShrink.AICglm.lm},
  depending on the class of the objects. 

  \code{modavg.shrink.gls} has been replaced by
  \code{modavgShrink.AICgls}. 

  \code{modavg.shrink.lme} has been replaced by
  \code{modavgShrink.AIClme}. 

  \code{modavg.shrink.mer} has been replaced by
  \code{modavgShrink.AICmer}. 

  \code{modavg.shrink.merMod} has been replaced by
  \code{modavgShrink.AICglmerMod} or
  \code{modavgShrink.AIClmerMod}, depending on the class of the
  objects.

  \code{modavg.shrink.mult} has been replaced by
  \code{modavgShrink.AICmultinom.nnet}. 

  \code{modavg.shrink.polr} has been replaced by
  \code{modavgShrink.AICpolr}. 

  \code{modavg.shrink.rlm} has been replaced by
  \code{modavgShrink.AICrlm.lm} 

  \code{modavg.shrink.unmarked} has been replaced by
  \code{modavgShrink.AICunmarkedFitOccu},
  \code{modavgShrink.AICunmarkedFitColExt}, 
  \code{modavgShrink.AICunmarkedFitOccuRN},
  \code{modavgShrink.AICunmarkedFitPCount},
  \code{modavgShrink.AICunmarkedFitPCO},
  \code{modavgShrink.AICunmarkedFitDS}, 
  \code{modavgShrink.AICunmarkedFitGDS},
  \code{modavgShrink.AICunmarkedFitOccuFP},
  \code{modavgShrink.AICunmarkedFitMPois},
  \code{modavgShrink.AICunmarkedFitGMM}, or
  \code{modavgShrink.AICunmarkedFitGPC}, depending on the class of
  the objects.

  \code{modavgpred} has been replaced by \code{modavgPred}.
  
  \code{modavgpred.glm} has been replaced by
  \code{modavgpred.AICglm.lm} or \code{modavgPred.AIClm},
  depending on the class of the objects.
  
  \code{modavgpred.gls} has been replaced by \code{modavgPred.AICgls}.  

  \code{modavgpred.lme} has been replaced by \code{modavgPred.AIClme}.

  \code{modavgpred.mer} has been replaced by \code{modavgPred.AICmer}.
  
  \code{modavgpred.merMod} has been replaced by
  \code{modavgpred.AICglmerMod} or \code{modavgPred.AIClmerMod},
  depending on the class of the objects.

  \code{modavgpred.rlm} has been replaced by
  \code{modavgPred.AICrlm.lm}. 

  \code{modavgpred.unmarked} has been replaced by
  \code{modavgPred.AICunmarkedFitOccu},
  \code{modavgPred.AICunmarkedFitColExt}, 
  \code{modavgPred.AICunmarkedFitOccuRN},
  \code{modavgPred.AICunmarkedFitPCount},
  \code{modavgPred.AICunmarkedFitPCO},
  \code{modavgPred.AICunmarkedFitDS}, 
  \code{modavgPred.AICunmarkedFitGDS},
  \code{modavgPred.AICunmarkedFitOccuFP},
  \code{modavgPred.AICunmarkedFitMPois},
  \code{modavgPred.AICunmarkedFitGMM}, or
  \code{modavgPred.AICunmarkedFitGPC}, depending on the class of
  the objects.

  \code{mult.comp} has been replaced by \code{multComp}.

  \code{predictSE.zip} has been replaced by \code{predictSE}.
}

\author{
Marc J. Mazerolle
}

\seealso{
\code{\link{aictab}}, \code{\link{confset}}, \code{\link{dictab}}, 
\code{\link{importance}}, \code{\link{evidence}},
\code{\link{extractLL}}, \code{\link{c_hat}},
\code{\link{modavg}}, \code{\link{modavgEffect}},
\code{\link{modavgShrink}}, \code{\link{modavgPred}},
\code{\link{multComp}}, \code{\link{predictSE}}
}

\keyword{models}